Refactor Funnel and Goal components to update import paths for analytics hooks. Consolidated imports from deprecated paths to new directory structure, enhancing code organization and maintainability.
