+++ client/src/app/[site]/funnels/components/FunnelForm.tsx
@@ -1,4 +1,3 @@
-import { FunnelResponse, FunnelStep } from "@/api/analytics/useGetFunnel";
 import { Time } from "@/components/DateSelector/types";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
@@ -16,9 +15,12 @@ import { ThreeDotLoader } from "../../../../components/Loaders";
 import { Filter } from "@rybbit/shared";
 import { FilterComponent } from "../../components/shared/Filters/FilterComponent";
 import { Funnel } from "./Funnel";
-import { SavedFunnel } from "../../../../api/analytics/useGetFunnels";
 import { Switch } from "../../../../components/ui/switch";
 import { Label } from "../../../../components/ui/label";
+import {
+  FunnelResponse,
+  FunnelStep,
+} from "../../../../api/analytics/funnels/useGetFunnel";
 
 interface FunnelFormProps {
   name: string;
